20251211_1

+++ src/controllers/videoController.js
@@ -1,45 +1,49 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
-  try {
-    const videos = await Video.find({});
-    console.log(videos);
-    return res.render("home", { pageTitle: "Home", videos });
-  } catch (error) {
-    console.log("server-error", { error });
-  }
+  const videos = await Video.find({})
+    .sort({ createdAt: "desc" })
+    .populate("owner");
+  return res.render("home", { pageTitle: "Home", videos });
 };
 
 export const watch = async (req, res) => {
-  const id = req.params.id;
-  const video = await Video.findById(id);
+  const { id } = req.params;
+  const video = await Video.findById(id).populate("owner");
   if (!video) {
-    return res.status(404).render("404", {
-      pageTitle: "비디오가 존재하지 않습니다.",
-    });
+    return res.render("404", { pageTitle: "Video not found." });
   }
   return res.render("watch", { pageTitle: video.title, video });
 };
 
 export const getEdit = async (req, res) => {
-  const id = req.params.id;
+  const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
-    return res.status(404).render("404", {
-      pageTitle: "비디오가 존재하지 않습니다.",
-    });
+    return res.status(404).render("404", { pageTitle: "Video not found." });
   }
-  return res.render("edit", { pageTitle: `에딧 : ${video.title}`, video });
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
-  const id = req.params.id;
+  const {
+    user: { _id },
+  } = req.session;
+  const { id } = req.params;
   const { title, description, hashtags } = req.body;
-  const video = await Video.findById(id);
+  const video = await Video.exists({ _id: id });
   if (!video) {
-    return res.status(404).render("404", {
-      pageTitle: "비디오가 존재하지 않습니다.",
-    });
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -50,28 +54,48 @@ export const postEdit = async (req, res) => {
 };
 
 export const getUpload = (req, res) => {
-  return res.render("upload", { pageTitle: "업로드" });
+  return res.render("upload", { pageTitle: "Upload Video" });
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      fileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
+    console.log(error);
     return res.status(400).render("upload", {
-      pageTitle: "업로드",
+      pageTitle: "Upload Video",
       errorMessage: error._message,
     });
   }
 };
 
 export const deleteVideo = async (req, res) => {
-  const id = req.params.id;
+  const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
@@ -82,9 +106,9 @@ export const search = async (req, res) => {
   if (keyword) {
     videos = await Video.find({
       title: {
-        $regex: new RegExp(keyword, "i"),
+        $regex: new RegExp(`${keyword}$`, "i"),
       },
-    });
+    }).populate("owner");
   }
-  return res.render("search", { pageTitle: "검색", videos });
+  return res.render("search", { pageTitle: "Search", videos });
 };
